Include go package (#65)
* Include the go_package option.

* Refactoring: Move service generation into separate file.

* Refactoring: Move messages generation into separate file.
