+++ examples/bookstore/bookstore.proto
@@ -8,6 +8,8 @@ import "google/protobuf/descriptor.proto";
 
 import "google/protobuf/empty.proto";
 
+option go_package = ".;bookstore";
+
 message Book {
   string author = 1;
