+++ generator/testfiles/goldstandard/parameters.proto
@@ -8,6 +8,8 @@ import "google/protobuf/descriptor.proto";
 
 import "google/protobuf/empty.proto";
 
+option go_package = ".;parameters";
+
 message Parameter1 {
   int64 param7 = 1;
 }
