+++ generator/testfiles/goldstandard/requestbodies.proto
@@ -8,6 +8,8 @@ import "google/protobuf/descriptor.proto";
 
 import "google/protobuf/empty.proto";
 
+option go_package = ".;requestbodies";
+
 message Person {
   int64 id = 1;
